변경
qna -> board 이름 변경
문의 페이지
마이 페이지
상세페이지

+++ teamproject_qna/src/main/java/com/example/teamproject_qna/repository/BoardRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.teamproject_qna.entity.BoardEntity;
 import com.example.teamproject_qna.entity.Member;
 
-public interface BoardRepository extends JpaRepository<BoardEntity, Integer>{
+public interface BoardRepository extends JpaRepository<BoardEntity, Long>{
 
 	List<BoardEntity> findByWriter(Member loginUser);
